cleaning the variables
* i_coltrans: check the initialization of an array linetransitions
* roverw: large part of this sbr was noted, unused variables were noted
  too
* roverw: unused input variables fr_line and l_dist
* connection_prop_model_grid: the variables for a parallelization were
  not deleted
* event_dist: tooRed was initialized to the value F
* types: I do not know, what does the forthcoming Warning mean
  Warning: Legacy Extension: Duplicate SAVE attribute specified at (1)
* rates_i: deleted (probably) the constants which are not used

+++ types.f90
@@ -78,7 +78,7 @@ SAVE
   END TYPE line_list
 
   TYPE ion_levels 
-     INTEGER(KIND=2), ALLOCATABLE    :: linetransitions(:), lineuptransitions(:)
+     INTEGER, ALLOCATABLE    :: linetransitions(:), lineuptransitions(:)
      DOUBLE PRECISION                :: exci_energy, stat_waight
      CHARACTER(LEN=30)               :: elconf
      LOGICAL                         :: phcrossform
